pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "rca-agent"
version = "0.1.0"
description = "AI Root Cause Analysis pipeline (LangGraph, FastAPI, Streamlit)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-ollama>=0.2.0",
    "langchain-anthropic>=0.3.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "streamlit>=1.40.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]
